valutatrade_hub/parser_service/config.py
import os
from dataclasses import dataclass
from types import MappingProxyType

from dotenv import load_dotenv

load_dotenv()


@dataclass
class ParserConfig:
    
    EXCHANGERATE_API_KEY: str = os.getenv("EXCHANGERATE_API_KEY", "")
    
    # Эндпоинты
    COINGECKO_URL: str = "https://api.coingecko.com/api/v3/simple/price"
    EXCHANGERATE_API_URL: str = "https://v6.exchangerate-api.com/v6"
    
    # Списки валют
    BASE_CURRENCY: str = "USD"
    FIAT_CURRENCIES: tuple = ("EUR", "GBP", "RUB")
    CRYPTO_CURRENCIES: tuple = ("BTC", "ETH", "SOL")
    CRYPTO_ID_MAP: dict = MappingProxyType({
        "BTC": "bitcoin",
        "ETH": "ethereum",
        "SOL": "solana",
    })
    
    # Пути к файлам
    RATES_FILE_PATH: str = "data/rates.json"
    HISTORY_FILE_PATH: str = "data/exchange_rates.json"
    
    # Сетевые параметры
    REQUEST_TIMEOUT: int = 10


config = ParserConfig()
