pyproject.toml
[project]
name = "finalproject-bazhenov-m25-555"
version = "0.1.0"
description = "A platform for tracking and simulating currency trading"
authors = [
    {name = "RRifen",email = "gorbaty.kit@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prettytable (>=3.17.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "valutatrade_hub" },
]

[dependency-groups]
dev = [
    "ruff (>=0.14.11,<0.15.0)"
]

[tool.poetry.scripts]
project = "main:main"

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["venv", ".venv", "dist", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = [] 

[tool.valutatrade]
data_dir = "data"
users_file = "users.json"
portfolios_file = "portfolios.json"
rates_file = "rates.json"
rates_ttl_seconds = 300
default_base_currency = "USD"
log_path = "logs/actions.log"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
